move out common code to single dummy dataset
